✅ test: one control per predicate clause (ADR-028 slice 4)
Slice 3's shape was a matched pair per predicate — shipped values satisfy it,
a broken input violates it. A sixth review pass found that is not enough: a
predicate with N independent clauses needs N controls, one failing each clause
alone. A single fixture that trips several at once is no evidence that any one
of them is separately enforced.

Two clauses had no evidence, and deleting either left every test green:

- on-accent base fill (>= 3.0 over `moss`). The existing control holds
  `inkOnAccent` white in dark, which fails both bars (1.911 / 2.126). #FBFBFB
  fails the base clause alone: 4.577 over `mossDark` clears the text bar,
  2.933 over `moss` misses the non-text one.
- quiet-line surface separability (>= 1.1 over the surface). The two existing
  controls are rejected on direction and on the ground reading. #323232 sits
  just above `nightBubble`, reads 1.327 against the ground and 1.061 against
  the surface.

The on-accent fixture is light-side by necessity, not preference:
`contrastRatio` depends only on Y, so the gap achievable between the two bars
is capped by the fills' own luminance ratio, and nightMossDark/nightMoss sit
1.113 apart — under the 4.5/3.0 = 1.5 the bars demand. No colour clears one
and misses the other in dark.

`mossBadgeQuietestControlRejectsALiftedInk` now injects its substituted ink
through `mossBadgeIsQuietest` instead of into the inner `isQuietestPair`.
Aimed at the inner predicate it skipped the wrapper entirely: collapsing the
wrapper to `return true`, or swapping its two branches, left every test green.
The added `dark: false` assertion is what makes a branch swap observable.

Each of the four mutations was run and read by which test reddened, not by the
exit code, with the anchor asserted first — the deletions redden exactly their
own new control, and the branch swap reddens both of the badge control's
assertions.

The predicates move to `DesignTokensTests+NightRemainderHelpers.swift`. The
file stood at 399 lines against swiftlint's 400-line cap (--strict makes it an
error), so the controls could not land without either trimming rationale or
moving structure, and trimming rationale to satisfy a line count is the wrong
trade — the same call `DesignTokens+NightStatePalette.swift` records.
